+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Fixed
+
+- Fixed home timeline marker appearing after publishing a post.
+
 ## [5.5.0] - 2026-06-10
 
 ### Added