+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Changed
+
+- Made `Account.is_group` field required.
+
 ## [5.6.0] - 2026-06-26
 
 ### Added