+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Added optional `max_id` query parameter to profile page URL ([#171](https://codeberg.org/silverpill/mitra-web/pulls/171)).
 - Added "Description" field to group creation form.
+- Added interface for deleting moderated groups.
 
 ### Changed