+++ CHANGELOG.md
@@ -20,6 +20,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Made `Account.is_group` field required.
 - Assign unique paths to tabs in group list.
 - Redirect to "Moderating" tab once a group has been created.
+- Assign unique paths to tabs in follow request list.
 
 ## [5.6.0] - 2026-06-26