+++ CHANGELOG.md
@@ -22,6 +22,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Assign unique paths to tabs in group list.
 - Redirect to "Moderating" tab once a group has been created.
 - Assign unique paths to tabs in follow request list.
+- Speed up loading of posts on profile page.
 
 ### Fixed