+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Converted "Group" badge into a link to group timeline.
 - Added "View group profile" item to post menu (admin only).
 - Show total number of voters in a poll.
+- Added pagination to subscriber list on profile page.
 
 ### Changed